refactor: 타인의 장바구니 품목을 수정 및 삭제할 수 없도록 변경

+++ src/main/java/cart/service/CartItemService.java
@@ -7,6 +7,7 @@ import cart.domain.product.Product;
 import cart.dto.CartItemQuantityUpdateRequest;
 import cart.dto.CartItemRequest;
 import cart.dto.CartItemResponse;
+import cart.exception.CartItemException;
 import cart.repository.cart.CartRepository;
 import cart.repository.product.ProductRepository;
 import org.springframework.stereotype.Service;
@@ -48,8 +49,9 @@ public class CartItemService {
     }
 
     public void updateQuantity(final Member member, final Long cartItemId, final CartItemQuantityUpdateRequest request) {
-        Cart cart = cartRepository.findCartByMemberId(member.getId());
         CartItem cartItem = cartRepository.findCartItemById(cartItemId);
+        Cart cart = cartRepository.findCartByMemberId(member.getId());
+        validateOwner(member, cart, cartItem);
 
         if (request.getQuantity() == 0) {
             cart.removeItem(cartItem);
@@ -63,10 +65,17 @@ public class CartItemService {
     }
 
     public void remove(final Member member, final Long cartItemId) {
-        Cart cart = cartRepository.findCartByMemberId(member.getId());
         CartItem cartItem = cartRepository.findCartItemById(cartItemId);
+        Cart cart = cartRepository.findCartByMemberId(member.getId());
+        validateOwner(member, cart, cartItem);
 
         cart.removeItem(cartItem);
         cartRepository.deleteCartItemById(cartItem.getId());
+    }
+
+    private void validateOwner(final Member member, final Cart cart, final CartItem cartItem) {
+        if (!cart.hasItem(cartItem)) {
+            throw new CartItemException.IllegalMember(cart, member);
+        }
     }
 }
